fix 64-bit warnings
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15471 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tad/cli/tad_cli_layer.c
@@ -115,7 +115,7 @@ cli_gen_bin_cb(csap_p csap_descr, int layer, const asn_value * tmpl_pdu,
                const csap_pkts_p  up_payload, csap_pkts_p pkts)
 {
     int rc;
-    int msg_len;
+    size_t msg_len;
 
     char *msg;
